+++ components/Util/numberToString.js
@@ -1,0 +1,5 @@
+const numberToString = (param) => {
+  return `${param}px`;
+};
+
+export default numberToString;