+++ styles/theme.js
@@ -1,5 +1,16 @@
+import {parseAlign} from './parsers/parseAlign';
+import {parseColorValue, parseBgColor, parseFontColor} from './parsers/parseColors';
+import {parseCustom, parseCustomDef} from './parsers/parseCustom';
+import {parseDimension} from './parsers/parseDimension';
+import {parseFontFamily} from './parsers/parseFontFamily';
+import {parseMargin} from './parsers/parseMargin';
+import {parsePadding} from './parsers/parsePadding';
+import {parseRadius} from './parsers/parseRadius';
+import {parseProps} from './parsers/parseRender';
+
 export const theme = {
   colors: {
+    activeColor: '#09B77B',
     primary1: '#063F8F',
     primary2: '#446FAB',
     primary3: '#829EC7',
@@ -38,7 +49,36 @@ export const theme = {
     error3: '#E37475',
   },
   fonts: {
-    fontPrimary: 'PT Sans',
-    fontSecondary: 'Nunito Sans'
+    primary: 'Work Sans',
+    secondary: 'Nunito Sans'
+  },
+  factors: {
+    dimensions: 10,
+    padding: 8,
+    margin: 8,
+  },
+  parseAlign,
+  parseBgColor,
+  parseCustom,
+  parseCustomDef,
+  parseDimension,
+  parseFontColor,
+  parseFontFamily,
+  parseMargin,
+  parsePadding,
+  parseProps,
+  parseRadius,
+  parseColorValue,
+  queries: {
+    xs: '360px', 
+    sm: '460px', 
+    md: '768px', 
+    lg: '1016px',
+    xl: '1280px'
+  },
+  radius: {
+    false: '0px',
+    default: '3px',
+    alternative: '5px'
   }
 };