+++ CHANGELOG.md
@@ -1,3 +1,28 @@
+# v1.4.5
+- Remove os exports de: 
+  export {default as SearchForm} from './Menus/SearchForm';
+  export {default as SearchMenu} from './Menus/Search';
+  export {default as SideMenu} from './Menus/SideMenu';
+  export {default as SideMenuItems} from './Menus/SideItems';
+- Novo componente SideMenu
+- Refatora componente Button
+- Novo componente AMP/Button
+# v1.4.4
+- Adiciona componente Blocks styled-components
+- Documentação do Blocks
+- Adiciona theme
+- Adiciona lógica de parseamento dos tokens dentro do theme
+# v1.4.2
+- Corrige funcionamento dos tamanhos de margin na função getFromProps
+- Export da pasta Util passou a ser import from prensa/utils
+# v1.4.1
+- Adiciona loading no componente Button
+- Adiciona função para parse do label no Field
+- Adiciona listener para teclas no Button
+# v1.4.0
+- Adiciona styled-components no Prensa
+- Adiciona componente Field
+
 # v1.3.8
 - Refatora template Gridnews (Prensa-Templates-Gridnews)
 - Refatora template Featured (Prensa-Templates-Featured)